Rewrite of Populator -> MongoPopulator.
Populator is a great gem for populating relational database tables when
using ActiveRecord models. This rewrite populates MongoDB collections,
and is framework agnostic---meant to be a standalone script.

This gem depends on the `mongo` gem, and the #populate method is added
to Mongo::Collection (instead of ActiveRecord::Base as in the original).
Though it started as a fork of https://github.com/ryanb/populator, it
has diverged significantly and does not have the same functionality,
so it will be a standalone repo and gem. The syntax is mostly the same
as Populator. There are two new methods for storing arrays in a doc,
and the method of relating records is different. Embedded documents are
not yet supported.

Squashed commit of the following:

commit 7dd9c44e447beac57a413048bce768d5869ad94f
Author: Benjamin Cullen-Kerney <ben.kerney@gmail.com>
Date:   Tue Sep 13 23:45:55 2011 -0700

    Eeeeenglishhhh

commit aa9959e25b6c262038916dae4a768374c40bfc91
Author: Benjamin Cullen-Kerney <ben.kerney@gmail.com>
Date:   Tue Sep 13 23:42:49 2011 -0700

    rename so requires aren't confusing for users, freshen up examples

commit 29fe798c82b0bd2dd4f010141c76b7e252a335aa
Author: Benjamin Cullen-Kerney <ben.kerney@gmail.com>
Date:   Tue Sep 13 23:19:31 2011 -0700

    Out, out

commit 493324077fd133ed3ecbe8d0603fa29bb974b7c7
Author: Benjamin Cullen-Kerney <ben.kerney@gmail.com>
Date:   Tue Sep 13 23:18:57 2011 -0700

    Require dependency

commit 560177141f1a801c6eaf7d009f21d6a83d4324e8
Author: Benjamin Cullen-Kerney <ben.kerney@gmail.com>
Date:   Tue Sep 13 23:09:31 2011 -0700

    mongo-populator relies on mongo, so install it

commit 04661c467089b3026d8dbec2e321e74836e365ab
Author: Benjamin Cullen-Kerney <ben.kerney@gmail.com>
Date:   Tue Sep 13 22:55:23 2011 -0700

    Straightforward application to MongoDB. Deleted inapplicable specs, got the others passing.

commit 35152dde2d9bb745d2a07a5662123bebf863be45
Author: Benjamin Cullen-Kerney <ben.kerney@gmail.com>
Date:   Tue Sep 13 15:42:01 2011 -0700

    Updated README